# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SchoolBox Offline API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /risk/classes/{classId}:
    get:
      operationId: getClassRiskScores
      tags: [risk]
      summary: Get risk scores for all students in a class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class risk scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRiskScore"
  /risk/students/{studentId}/history:
    get:
      operationId: getStudentRiskHistory
      tags: [risk]
      summary: Get risk score history for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk score history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskScoreHistory"
  /risk/alerts:
    get:
      operationId: listRiskAlerts
      tags: [risk]
      summary: List risk alerts
      parameters:
        - name: acknowledged
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Risk alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskAlert"
  /risk/alerts/{id}/acknowledge:
    patch:
      operationId: acknowledgeRiskAlert
      tags: [risk]
      summary: Acknowledge risk alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAlert"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, sent, failed, null]
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{studentId}/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send manual notification to student parent
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationBody"
      responses:
        "200":
          description: Notification queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  schemas:
...
    StudentRiskScore:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        score:
          type: number
        tier:
          type: string
          enum: [low, medium, high, critical]
        explanation:
          type: array
          items:
            type: string
        trend:
          type: string
          enum: [up, down, stable]
        consecutiveAbsences:
          type: integer
        absenceRate30d:
          type: number
        computedAt:
          type: string
      required:
        - studentId
        - studentName
        - classId
        - score
        - tier
        - explanation
        - trend
        - consecutiveAbsences
        - absenceRate30d
        - computedAt
    RiskScoreHistory:
      type: object
      properties:
        date:
          type: string
        score:
          type: number
        tier:
          type: string
          enum: [low, medium, high, critical]
      required:
        - date
        - score
        - tier
    RiskAlert:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        className:
          type: ["string", "null"]
        tier:
          type: string
          enum: [low, medium, high, critical]
        score:
          type: number
        alertType:
          type: string
        triggeredAt:
          type: string
        acknowledgedAt:
          type: ["string", "null"]
        acknowledgedBy:
          type: ["string", "null"]
      required:
        - id
        - studentId
        - studentName
        - classId
        - tier
        - score
        - alertType
        - triggeredAt
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        templateKey:
          type: string
        lang:
          type: string
          enum: [fr, ar]
        status:
          type: string
          enum: [pending, sent, failed]
        transportUsed:
          type: ["string", "null"]
        retries:
          type: integer
        sentAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - studentId
        - templateKey
        - lang
        - status
        - retries
        - createdAt
    SendNotificationBody:
      type: object
      properties:
        templateKey:
          type: string
          enum: [absence_alert, risk_alert, quiz_failure]
        lang:
          type: string
          enum: [fr, ar]
      required:
        - templateKey
        - lang